refactor: extract results from search component
